feat(persist-user): refactor into model and builder packages
